refactor: rename class with print feat
